Add files via upload
Articulation point/critical points find those vertex(s) that are crucial for the graph. If that vertex is removed from the graph, all vertices cannot be traversed. 
Maintain an array "discovery" and another array called "low". Use DFS to traverse the graph, and once a call is made to the function DFS::traverse(int v, int parent), update discovery and low with the discovery counter. this defines basically the time taken to reach this vertex from the parent vertex. Use iterator to call all the vertex(s), Now, for every Vertex(excluding the starting vertex), it has a parent vertex from which it was called, a neighbour( pseudo parent) which could have called if not called by parent(visited == true) and child(ren). The purpose of traversing each node through bfs is to update the array "low". if through iteration a vertex is neighbour, update low. If it is child, call dfs and then after returning update the low again. if visiing time of vertex v is less than low of vertex child then it means v is a crtical connection since child 's low (Which should be updated to a smalled value if there were neighbours ) didnt have neighbours hence the only way it was updated / inserted was through parent.